requirements.txt
google-api-python-client==2.131.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
requests==2.32.3
Pillow==10.4.0
feedparser==6.0.11
